chore: reorganize CI workflows and python files
