--- .github/workflows/benchmark.yml
@@ -1,19 +1,0 @@
-name: Benchmarks
-
-on:
-  push:
-    branches:
-      - master
-      - dev
-
-jobs:
-  benchmark:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Update
-        run: rustup update
-
-      - name: run
-        run: cargo bench
