--- .github/workflows/ci.yaml
@@ -1,16 +1,0 @@
-name: CI
-
-on:
-  push
-
-jobs:
-  pre-commit:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Update rust toolchain
-        run: rustup update
-      - uses: actions/setup-python@v3
-        with:
-          python-version: 3.9
-      - uses: pre-commit/action@v3.0.0
