--- .github/workflows/clippy.yml
@@ -1,19 +1,0 @@
-name: Clippy
-
-on: [push]
-
-jobs:
-  clippy:
-    name: Clippy
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Update
-        run: rustup update
-
-      - name: Install
-        run: rustup component add clippy
-
-      - name: Run
-        run: cargo clippy -- -D warnings
