+++ .github/workflows/release.yml
@@ -4,52 +4,30 @@ on:
   push:
     branches:
       - master
+      - chore/poetry
 
 jobs:
   cargo:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-
       - name: Update
         run: rustup update
-
-      - name: Test
-        run: cargo test --release
-
-      - name: Build
-        run: docker build -t release .
-
       - name: Publish
-        # To make this actually run, add -w yes
-        run: >
-          docker run
-          --env GH_TOKEN="${{ secrets.GITHUB_TOKEN }}"
-          --env CARGO_TOKEN="${{ secrets.CARGO_TOKEN }}"
-          --env GIT_COMMITTER_NAME="URI ABD"
-          --env GIT_COMMITTER_EMAIL="info@uri-abd.org"
-          release /bin/bash -c "cargo install semantic-rs && semantic-rs"
+        run: cargo publish --token ${{ secrets.CARGO_SECRET }}
 
   pypi:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-
-      - name: Python
-        uses: actions/setup-python@v1
+      - uses: actions/setup-python@v2
         with:
           python-version: 3.9
-
       - name: Dependencies
         run: |
-          python -m pip install --upgrade pip setuptools wheel
-          pip install twine toml
-          pip install .
-
+          pip install --user poetry && poetry install
       - name: Publish
         env:
-          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-        run: |
-          python setup.py sdist bdist_wheel
-          twine upload dist/*
+          PYPI_USERNAME: ${{ secrets.PYPI_USERNAME }}
+          PYPI_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+        run: poetry publish --build --username $PYPI_USERNAME --password $PYPI_PASSWORD
