+++ .github/workflows/tests.yml
@@ -3,6 +3,17 @@ name: Tests
 on: [push]
 
 jobs:
+  pre-commit:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Update rust toolchain
+        run: rustup update
+      - uses: actions/setup-python@v3
+        with:
+          python-version: 3.9
+      - uses: pre-commit/action@v3.0.0
+
   cargo:
     runs-on: ubuntu-latest
     steps:
@@ -15,11 +26,14 @@ jobs:
       run: cargo test --release
 
   unittest:
-    runs-on: python:3.9
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+      - uses: actions/setup-python@v3
+        with:
+          python-version: 3.9
       - name: Dependencies
         run: |
           pip install --user poetry && poetry install
       - name: Run
-        run: python -m unittest discover
+        run: poetry run python -m unittest discover python
