+++ .pre-commit-config.yaml
@@ -6,6 +6,7 @@ repos:
   hooks:
     - id: debug-statements
     - id: trailing-whitespace
+      exclude: '.bumpversion.cfg'
     - id: end-of-file-fixer
     - id: check-yaml
     - id: check-toml
