+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "clam"
-version = "0.9.0"
+name = "abd-clam"
+version = "0.9.0-dev5"
 authors = [
     "Najib Ishaq <najib_ishaq@zoho.com>",
     "Tom Howard <info@tomhoward.codes>",
@@ -11,11 +11,13 @@ authors = [
 ]
 edition = "2021"
 rust-version = "1.68"
+description = "Clustered Learning of Approximate Manifolds"
+license = "MIT"
 
 # Publishing lock.
 # When you're ready to start publishing, set publish to true
 # and update the release github action.
-publish = false
+publish = true
 
 [profile.test]
 debug = 1
