+++ python/abd_clam/core/graph_criteria.py
