+++ python/abd_clam/utils/constants.py
